common/blast-types/src/test/kotlin/org/veupathdb/lib/blast/common/fields/NegativeIPGListTest.kt
package org.veupathdb.lib.blast.common.fields

import org.junit.jupiter.api.Assertions.*
import org.junit.jupiter.api.DisplayName
import org.junit.jupiter.api.Nested
import org.junit.jupiter.api.Test
import org.junit.jupiter.api.assertThrows
import org.veupathdb.lib.blast.common.FlagNegativeIPGList
import org.veupathdb.lib.jackson.Json

@DisplayName(FlagNegativeIPGList)
internal class NegativeIPGListTest {

  @Nested
  @DisplayName("ParseNegativeIPGList()")
  inner class Parse {

    @Nested
    @DisplayName("when given a JSON object containing a non-text value")
    inner class NonString {

      @Test
      @DisplayName("throws an exception")
      fun t1() {
        assertThrows<IllegalArgumentException> {
          ParseNegativeIPGList(Json.new { put(FlagNegativeIPGList, 40) })
        }
      }
    }

    @Nested
    @DisplayName("when given a JSON object with no db file value")
    inner class Absent {

      @Test
      @DisplayName("returns a defaulted IPGList instance")
      fun t1() {
        assertTrue(ParseNegativeIPGList(Json.new {}).isDefault)
      }
    }

    @Nested
    @DisplayName("when given a JSON object containing a text value")
    inner class Present {

      @Test
      @DisplayName("returns a IPGList instance wrapping")
      fun t1() {
        val inp = Json.newObject { put(FlagNegativeIPGList, "bunny") }
        assertEquals("bunny", ParseNegativeIPGList(inp).file)
      }
    }
  }

  @Nested
  @DisplayName("IPGList()")
  inner class Type {

    @Nested
    @DisplayName("when constructed with no arguments")
    inner class NoArg {

      @Test
      @DisplayName("default == true")
      fun t1() {
        assertTrue(NegativeIPGList().isDefault)
      }

      @Test
      @DisplayName("appendJson() does nothing")
      fun t2() {
        val inp = Json.newObject()
        NegativeIPGList().appendJson(inp)
        assertEquals(0, inp.size())
      }

      @Test
      @DisplayName("appendCliSegment() does nothing")
      fun t3() {
        val inp = StringBuilder(0)
        NegativeIPGList().appendCliSegment(inp)
        assertEquals("", inp.toString())
      }

      @Test
      @DisplayName("appendCliParts() does nothing")
      fun t4() {
        val inp = ArrayList<String>(0)
        NegativeIPGList().appendCliParts(inp)
        assertTrue(inp.isEmpty())
      }
    }

    @Nested
    @DisplayName("when constructed with an argument")
    inner class Arg {

      @Test
      @DisplayName("isDefault == false")
      fun t1() {
        assertFalse(NegativeIPGList("bye.db").isDefault)
      }

      @Test
      @DisplayName("appendJson() appends the flag to the JSON object input.")
      fun t2() {
        val inp = Json.newObject()
        NegativeIPGList("goop.txt").appendJson(inp)
        assertEquals("""{"$FlagNegativeIPGList":"goop.txt"}""", inp.toString())
      }

      @Test
      @DisplayName("appendCliSegment() appends the flag to the StringBuilder input.")
      fun t3() {
        val inp = StringBuilder(32)
        NegativeIPGList("nope.nope.nope").appendCliSegment(inp)
        assertEquals(" $FlagNegativeIPGList 'nope.nope.nope'", inp.toString())
      }

      @Test
      @DisplayName("appendCliParts() appends the flag to the List input.")
      fun t4() {
        val inp = ArrayList<String>(2)
        NegativeIPGList("yup.yup.yup").appendCliParts(inp)
        assertEquals(2, inp.size)
        assertEquals(FlagNegativeIPGList, inp[0])
        assertEquals("yup.yup.yup", inp[1])
      }
    }
  }
  

}